+++ markmeld/melder.py
@@ -1,6 +1,5 @@
 import datetime
 import frontmatter
-import glob
 import jinja2
 import os
 import re
@@ -20,8 +19,6 @@ from ubiquerg import is_url
 
 from .const import PKG_NAME
 from .exceptions import *
-
-from .utilities import format_command, recursive_get, run_cmd, make_abspath
 from .utilities import *
 
 MD_FILES_KEY = "md_files"
@@ -32,29 +29,9 @@ YAML_GLOBS_UNKEYED_KEY = "yaml_globs_unkeyed"
 
 _LOGGER = getLogger(PKG_NAME)
 
-# Embed these in the package?
-mm_targets = {
-    "figs": "/home/nsheff/code/sciquill/bin/build-pdfs fig",
-    "figs_png": "/home/nsheff/code/sciquill/bin/buildfigs fig/*.svg",
-    "yaml_refs": "jabref -n --exportMatches 'groups=shefflab',reflists/ncs_papers.yaml,my_yaml i ${HOME}/code/papers/sheffield.bib",
-    "pubs": "",
-    "split": "/home/nsheff/code/sciquill/bin/splitsupl {combined} {primary} {appendix}",
-}
-
-tpl_generic = """{% if data.metadata_yaml is defined %}---
-{{ data.metadata_yaml }}
----{% endif %}
-
-{{ data.content }}"""
-
-
 tpl_generic = """
 {{ _global_frontmatter.fenced}}{{ content }}
 """
-
-# tpl_generic = """
-# {{ data }}
-# """
 
 
 @pass_environment
@@ -85,134 +62,33 @@ FILTERS["date"] = datetimeformat
 # Add custom reference extraction filter
 FILTERS["extract_refs"] = extract_refs
 
-
 # m = extract_refs("abc; hello @test;me @second one; and finally @three")
 # m
 
 
-def populate_data_yaml(cfg, data):
-    # Load up yaml data
-    if "data_yaml" not in cfg:
-        return data
-    _LOGGER.info(f"MM | Populating yaml data...")
-    for d in cfg["data_yaml"]:
-        _LOGGER.info(f"MM | {d}")
-        dabs = make_abspath(d, cfg["_cfg_file_path"])
+def get_frontmatter_formats(frontmatter):
+    """
+    Given a dictionary of content, return 3 versions of it: the dict, a yaml dumped version, and a fenced yaml dumped
 
-        with open(dabs, "r") as f:
-            data.update(yaml.load(f, Loader=yaml.SafeLoader))
+    @param dict frontmatter A dict representing some yaml frontmatter for a md file
+    """
+    if len(frontmatter) == 0:
+        frontmatter_raw = ""
+        frontmatter_fenced = ""
+    else:
+        frontmatter_raw = yaml.dump(frontmatter)
+        frontmatter_fenced = f"---\n{frontmatter_raw}---\n"
 
-    return data
-
-
-# For itemized yaml this time...
-def populate_data_yaml_keyed(cfg, data):
-    # Load up yaml data
-    if "data_yaml_keyed" not in cfg:
-        return data
-    _LOGGER.info(f"MM | Populating keyed yaml data...")
-    for k, v in cfg["data_yaml_keyed"].items():
-        _LOGGER.info(f"MM | --> {k}: {v}")
-        with open(v, "r") as f:
-            yaml_dict = yaml.load(f, Loader=yaml.SafeLoader)
-            _LOGGER.debug(yaml_dict)
-            data[k] = yaml_dict
-            data["yaml"][k] = yaml_dict
-            data["raw"][k] = yaml.dump(yaml_dict)
-            # 2022-08-12 Original way, this doesn't work in the case that data[k] is a list
-            # (if the yaml file is an array, not an object)
-            # So I changed it put the raw value under ["raw"][k] instead of [k]["raw"]
-            # data[k]["raw"] = yaml.dump(yaml_dict)
-    return data
+    return {
+        "raw": frontmatter_raw,
+        "fenced": frontmatter_fenced,
+        "dict": frontmatter,
+    }
 
 
-def populate_data_md(cfg, data):
-    # Load up markdown data
-    if "data_md" not in cfg:
-        return data
-
-    if "md" not in data:
-        data["md"] = {}
-
-    _LOGGER.info(f"MM | Populating md data...")
-    for k, v in cfg["data_md"].items():
-        _LOGGER.info(f"MM | --> {k}: {v}")
-        if not v:
-            data[k] = v
-            continue
-        if is_url(v):  # Do url stuff
-            import requests
-
-            response = requests.get(v)
-            p = frontmatter.loads(response.text)
-        else:
-            vabs = make_abspath(v, cfg["_cfg_file_path"])
-            if os.path.exists(vabs):
-                p = frontmatter.load(vabs)
-            else:
-                _LOGGER.warning(f"Skipping file that does not exist: {vabs}")
-                data[k] = {}
-                data["md"][k] = {}
-                data[k]["all"] = ""
-                continue
-        data[k] = p.__dict__
-        del data[k]["handler"]
-        data["md"][k] = data[k]
-        data[k]["all"] = frontmatter.dumps(p)
-        _LOGGER.debug(data[k])
-        if len(p.metadata) > 0:
-            data[k]["metadata_yaml"] = yaml.dump(p.metadata)
-
-    return data
-
-
-def populate_data_md_globs(cfg, data):
-    import glob
-
-    if "data_md_globs" not in cfg:
-        return data
-    _LOGGER.info(f"MM | Populating md data globs...")
-    for folder in cfg["data_md_globs"]:
-        files = glob.glob(folder)
-        _LOGGER.info(files)
-        for file in files:
-            basename = os.path.basename(file)
-            dirname = os.path.dirname(file)
-            splitext = os.path.splitext(basename)
-            k = splitext[0]
-            ext = splitext[1]
-            _LOGGER.info(f"MM | [key:value] {k}:{file}")
-            p = frontmatter.load(file)
-            data[k] = p.__dict__
-            del data[k]["handler"]
-            data["md"][k] = p.__dict__
-            del data["md"][k]["handler"]
-            data[k]["all"] = frontmatter.dumps(p)
-            data[k]["path"] = file
-            data[k]["ext"] = ext
-            if len(p.metadata) > 0:
-                data[k]["metadata_yaml"] = yaml.dump(p.metadata)
-    return data
-
-
-def resolve_globs(globs, cfg_path):
-    return_items = {}
-    if not globs:
-        return return_items
-    for item in globs:
-        path = os.path.join(os.path.dirname(cfg_path), item)
-        _LOGGER.info(f"MM | Glob path: {path}")
-        files = glob.glob(path)
-        for file in files:
-            k = os.path.splitext(os.path.basename(file))[0]
-            _LOGGER.info(f"MM | [key:value] {k}:{file}")
-            return_items[k] = file
-    return return_items
-
-
-def process_data_block(data_block, filepath):
+def process_data(data_block, filepath):
     _LOGGER.info(f"MM | Processing data block...")
-    data = {"_raw": {}}  # Initialize return value
+    data = {"_raw": {}, "_md": {}, "_yaml": {}}  # Initialize return value
     frontmatter_temp = {}
     local_frontmatter_temp = {}
     vars_temp = {}
@@ -223,13 +99,13 @@ def process_data_block(data_block, filepath):
 
     if MD_GLOBS_KEY in data_block and data_block[MD_GLOBS_KEY]:
         _LOGGER.info(f"MM | Populating md data globs...")
-        md_files.update(resolve_globs(data_block[MD_GLOBS_KEY], filepath))
+        md_files.update(globs_to_dict(data_block[MD_GLOBS_KEY], filepath))
     if YAML_GLOBS_KEY in data_block and data_block[YAML_GLOBS_KEY]:
         _LOGGER.info(f"MM | Populating yaml data globs...")
-        yaml_files.update(resolve_globs(data_block[YAML_GLOBS_KEY], filepath))
+        yaml_files.update(globs_to_dict(data_block[YAML_GLOBS_KEY], filepath))
     if YAML_GLOBS_UNKEYED_KEY in data_block and data_block[YAML_GLOBS_UNKEYED_KEY]:
         _LOGGER.info(f"MM | Populating unkeyed yaml globs...")
-        tmp_files = resolve_globs(data_block[YAML_GLOBS_UNKEYED_KEY], filepath)
+        tmp_files = globs_to_dict(data_block[YAML_GLOBS_UNKEYED_KEY], filepath)
         yaml_files.update(tmp_files)
         unkeyed_yaml_files = tmp_files
     if MD_FILES_KEY in data_block and data_block[MD_FILES_KEY]:
@@ -249,10 +125,12 @@ def process_data_block(data_block, filepath):
                 # data[k] = yaml_dict
                 if k in unkeyed_yaml_files:
                     data.update(yaml_dict)
+                    data["_yaml"].update(yaml_dict)
                 else:
                     data[k] = yaml_dict
+                    data["_yaml"][k] = yaml_dict
+                    vars_temp[k] = yaml_dict
                 data["_raw"][k] = yaml.dump(yaml_dict)
-                vars_temp.update(yaml_dict)
                 if k[:11] == "frontmatter":
                     frontmatter_temp.update(yaml_dict)
 
@@ -276,6 +154,7 @@ def process_data_block(data_block, filepath):
                 data["_raw"][k] = {}
                 continue
         data[k] = p.content
+        data["_md"][k] = p.content
         frontmatter_temp.update(p.metadata)
         local_frontmatter_temp[k] = p.metadata
         data["_raw"][k] = frontmatter.dumps(p)
@@ -295,19 +174,6 @@ def process_data_block(data_block, filepath):
                 frontmatter_temp.update({k[12:]: v})
 
     # vars_raw = yaml.dump(vars_temp)
-    def get_frontmatter_formats(frontmatter_temp):
-        if len(frontmatter_temp) == 0:
-            frontmatter_raw = ""
-            frontmatter_fenced = ""
-        else:
-            frontmatter_raw = yaml.dump(frontmatter_temp)
-            frontmatter_fenced = f"---\n{frontmatter_raw}---\n"
-
-        return {
-            "raw": frontmatter_raw,
-            "fenced": frontmatter_fenced,
-            "dict": frontmatter_temp,
-        }
 
     # Global vars behaves exactly like global frontmatter, except:
     # 1. It's all variables, not just those marked with frontmatter_*.
@@ -315,9 +181,8 @@ def process_data_block(data_block, filepath):
     # and excludes markdown content... Is that useful?
     data["_global_vars"] = vars_temp
 
-    # Integrated, global frontmatter -- combines all frontmatter from .md files,
-    # Plus any yaml data indexed with frontmatter_*,
-    # Plus any variables indexed with frontmatter_* -- in that priority order.
+    # Integrated, global frontmatter from 3 sources, in order:
+    # .md frontmatter, yaml data named frontmatter_*, variables named frontmatter_*
     data["_global_frontmatter"] = get_frontmatter_formats(frontmatter_temp)
 
     # Local frontmatter (per markdown file)
@@ -375,12 +240,13 @@ class Target(object):
     def __init__(self, root_cfg={}, target_name=None, vardata=None):
         self.root_cfg = root_cfg
         self.target_name = target_name
-        target_meta = {}
-        target_meta.update(self.root_cfg)
-        target_meta["_now"] = date.today().strftime("%s")
-        target_meta["_today"] = date.today().strftime("%Y-%m-%d")
-        target_meta["today"] = target_meta["_today"]  # TODO: Remove this
-        target_meta["now"] = target_meta["_now"]  # TODO: Remove this
+        meta = {}
+        # Old way would update based on root config:
+        # meta.update(self.root_cfg)
+        meta["_now"] = date.today().strftime("%s")
+        meta["_today"] = date.today().strftime("%Y-%m-%d")
+        meta["today"] = meta["_today"]  # TODO: Remove this
+        meta["now"] = meta["_now"]  # TODO: Remove this
 
         # Since a target has available to it all the variables in the _markemeld.yaml
         # config file, we start from there, then make a few changes:
@@ -397,35 +263,67 @@ class Target(object):
             if target_name not in list(root_cfg["targets"].keys()):
                 _LOGGER.error(f"target {target_name} not found")
                 raise TargetError(f"Target {target_name} not found")
-            target_meta = deep_update(target_meta, root_cfg["targets"][target_name])
+            if "inherit_from" in root_cfg["targets"][target_name]:
+                inherit_from = root_cfg["targets"][target_name]["inherit_from"]
+                if type(inherit_from) is not list:
+                    inherit_from = [inherit_from]
+                for base_target in inherit_from:
+                    _LOGGER.info(f"Loading from base target: {base_target}")
+                    meta = deep_update(meta, root_cfg["targets"][base_target])
+            meta = deep_update(meta, root_cfg["targets"][target_name])
             _LOGGER.debug(f'Config for this target: {root_cfg["targets"][target_name]}')
 
-        del target_meta["targets"]
-        if "version" in target_meta:
-            del target_meta["version"]
+        # del meta["targets"]
+        meta["_cfg_file_path"] = root_cfg["_cfg_file_path"]
+        if "version" in meta:
+            del meta["version"]
 
         if vardata:
             cli_vars = {y[0]: y[1] for y in [x.split("=") for x in vardata]}
-            target_meta.update(cli_vars)
+            meta.update(cli_vars)
         else:
             cli_vars = {}
 
-        if not "command" in target_meta:
+        if not "command" in meta:
             # Generally, user should provide a `command`, but for simple default cases,
             # we can just route through pandoc as a default command.
             options_array = []
-            if "latex_template" in target_meta:
-                options_array.append("--template {latex_template}")
-            if "output_file" in target_meta:
+            if "latex_template" in meta:
+                options_array.append('--template "{latex_template}"')
+            if "output_file" in meta:
                 options_array.append('--output "{output_file}"')
             options = " ".join(options_array)
-            target_meta["command"] = f"pandoc {options}"
+            meta["command"] = f"pandoc {options}"
 
-        _LOGGER.debug(f"target_meta: {target_meta}")
-        self.target_meta = target_meta
-        _LOGGER.debug(f"MM | Config file path: {self.target_meta['_cfg_file_path']}")
-        if "output_file" in self.target_meta:
-            _LOGGER.info(f"MM | Output file: {self.target_meta['output_file']}")
+        _LOGGER.debug(f"meta: {meta}")
+        self.meta = meta
+        _LOGGER.debug(f"MM | Config file path: {self.meta['_cfg_file_path']}")
+        if "output_file" in self.meta:
+            _LOGGER.info(f"MM | Output file: {self.meta['output_file']}")
+
+
+def build_side_targets(tgt, side_list_key="prebuild"):
+    """
+    Builds side targets for a target, which are prebuilds or postbuilds.
+
+    Side targets accompany a target, are built either before (prebuild)
+    or after (postbuild) a main target.
+
+    @param tgt Target The main target to build
+    @param side_list_key Iterable[str] The key of the target that contains a list of side targets. e.g. "prebuild" or "postbuild"
+    """
+    if side_list_key in tgt.meta:
+        _LOGGER.info(f"MM | Run {side_list_key} for target: {tgt.target_name}")
+        for side_tgt in tgt.meta[side_list_key]:
+            _LOGGER.info(f"MM | {side_list_key} target: {side_tgt}")
+            if side_tgt in self.cfg["targets"]:
+                self.build_target(side_tgt)
+            else:
+                _LOGGER.warning(
+                    f"MM | No target called {side_tgt}, requested prebuild by target {tgt}."
+                )
+                return False
+    return True
 
 
 class MarkdownMelder(object):
@@ -440,8 +338,8 @@ class MarkdownMelder(object):
     def open_target(self, target_name):
         tgt = Target(self.cfg, target_name)
 
-        if tgt.target_meta["output_file"] and not "stopopen" in tgt.target_meta:
-            return tgt.target_meta["output_file"]
+        if tgt.meta["output_file"] and not "stopopen" in tgt.meta:
+            return tgt.meta["output_file"]
         else:
             return False
 
@@ -451,75 +349,61 @@ class MarkdownMelder(object):
         _LOGGER.info(f"MM | Building target: {tgt.target_name}")
 
         # First, run any pre-builds
-        if "prebuild" in tgt.target_meta:
-            _LOGGER.info(f"MM | Run prebuilds for target: {tgt.target_name}")
-            for pretgt in tgt.target_meta["prebuild"]:
-                _LOGGER.info(f"MM | Prebuild target: {pretgt}")
-                if pretgt in self.cfg["targets"]:
-                    self.build_target(pretgt)
-                else:
-                    _LOGGER.warning(
-                        f"MM | No target called {pretgt}, requested prebuild by target {tgt}."
-                    )
-                    return False
+        if not build_side_targets(tgt, "prebuild"):
+            return False
 
         # Next, meld the inputs. This can be time-consuming, it reads data to populate variables
-        melded_input = self.meld_inputs(tgt)
-
-        if "loop" in tgt.target_meta:
-            return self.build_target_in_loop(tgt, melded_input, print_only, vardump)
+        tgt.melded_input = self.meld_inputs(tgt)
+        _LOGGER.debug(f"Melded input: {tgt.melded_input}")
+        if "loop" in tgt.meta:
+            return self.build_target_in_loop(tgt, print_only, vardump)
 
         # Run command...
-        return self.run_command_for_target(tgt, melded_input, print_only, vardump)
+        result = self.run_command_for_target(tgt, print_only, vardump)
 
-    def run_command_for_target(self, tgt, melded_input, print_only, vardump=False):
-        _LOGGER.info(f"File path for this target: {tgt.target_meta['_filepath']}")
-        if "type" in tgt.target_meta and tgt.target_meta["type"] == "raw":
+        # Finally, run any postbuilds
+        if not build_side_targets(tgt, "postbuild"):
+            return False
+
+        return result
+
+    def run_command_for_target(self, tgt, print_only, vardump=False):
+
+        _LOGGER.info(f"File path for this target: {tgt.meta['_filepath']}")
+        if "type" in tgt.meta and tgt.meta["type"] == "raw":
             # Raw = No subprocess stdin printing. (so, it doesn't render anything)
             cmd_fmt = format_command(tgt)
             tgt.melded_output = None
-            tgt.returncode = run_cmd(cmd_fmt, None, tgt.target_meta["_filepath"])
-        elif "type" in tgt.target_meta and tgt.target_meta["type"] == "meta":
+            tgt.returncode = run_cmd(cmd_fmt, None, tgt.meta["_filepath"])
+        elif "type" in tgt.meta and tgt.meta["type"] == "meta":
             # Meta = No command, it's a meta-target used for prebuilds or something else
             tgt.melded_output = None
             tgt.returncode = 0
         elif print_only:
             # Case 2: print_only means just render but run no command.
             # return print(tpl.render(data))  # one time
-            tgt.melded_output = self.render_template(melded_input, tgt)
+            tgt.melded_output = self.render_template(tgt.melded_input, tgt)
             tgt.returncode = 0
         elif vardump:
-            tgt.melded_output = melded_input
+            tgt.melded_output = tgt.melded_input
             tgt.returncode = 0
-        elif tgt.target_meta["command"]:
+        elif tgt.meta["command"]:
             cmd_fmt = format_command(tgt)
             _LOGGER.debug(cmd_fmt)
-            if (
-                "recursive_render" in tgt.target_meta
-                and not tgt.target_meta["recursive_render"]
-            ):
-                melded_output = self.render_template(
-                    melded_input, tgt, double=False
-                ).encode()
-            else:
-                # Recursive rendering allows your template to include variables
-                melded_output = self.render_template(
-                    melded_input, tgt, double=True
-                ).encode()
-            tgt.melded_output = melded_output
-            _LOGGER.debug(tgt.target_meta)
+            tgt.melded_output = self.render_template(tgt.melded_input, tgt)
             tgt.returncode = run_cmd(
-                cmd_fmt, melded_output, tgt.target_meta["_filepath"]
+                cmd_fmt, tgt.melded_output.encode(), tgt.meta["_filepath"]
             )
         return tgt
 
-    def build_target_in_loop(self, tgt, melded_input, print_only=False, vardump=False):
+    def build_target_in_loop(self, tgt, print_only=False, vardump=False):
         #  Process each iteration of the loop
-        loop_dat = recursive_get(
-            melded_input, tgt.target_meta["loop"]["loop_data"].split(".")
-        )
-        _LOGGER.debug(loop_dat)
-        _LOGGER.debug(tgt.root_cfg)
+        melded_input = tgt.melded_input
+        loop_data_var = tgt.meta["loop"]["loop_data"].split(".")
+        _LOGGER.debug(f"Retrieve loop data variable named {loop_data_var}")
+        loop_dat = recursive_get(melded_input, loop_data_var)
+        _LOGGER.debug(f"Loop dat: {loop_dat}")
+        _LOGGER.debug(f"Target melded_input: {tgt.melded_input}")
         n = len(loop_dat)
         _LOGGER.info(f"Loop found: {n} elements.")
         _LOGGER.debug(loop_dat)
@@ -529,48 +413,37 @@ class MarkdownMelder(object):
             loop_var_value = loop_dat[i]
             melded_input_copy = deepcopy(melded_input)
             tgt_copy = deepcopy(tgt)
-            var = tgt_copy.target_meta["loop"]["assign_to"]
+            var = tgt_copy.meta["loop"]["assign_to"]
             _LOGGER.info(f"{var}: {loop_var_value}")
-            melded_input_copy.update({var: loop_var_value})
-            tgt_copy.target_meta.update({var: loop_var_value})
-            _LOGGER.debug(tgt_copy.target_meta)
+            tgt_copy.melded_input.update({var: loop_var_value})
+            tgt_copy.meta.update({var: loop_var_value})
+            _LOGGER.debug(tgt_copy.meta)
             # _LOGGER.debug(cmd_data)
             rendered_in = self.render_template(
-                melded_input_copy, tgt_copy, double=False
+                tgt_copy.melded_input, tgt_copy, double=False
             ).encode()
             return_target_objects[i] = self.run_command_for_target(
-                tgt_copy, melded_input_copy, print_only, vardump
+                tgt_copy, print_only, vardump
             )
 
         return return_target_objects
 
-    def meld_inputs(self, target):
-        data_copy = deepcopy(target.root_cfg)
-        data_copy.update(target.target_meta)
+    def meld_inputs(self, tgt):
+        # data_copy = deepcopy(tgt.root_cfg)
+        # data_copy.update(tgt.meta)
+        data_copy = deepcopy(tgt.meta)
 
-        if not "version" in target.root_cfg or target.root_cfg["version"] < 1:
-            _LOGGER.info("MM | Processing config version 0...")
-            data_copy["yaml"] = {}
-            data_copy["raw"] = {}
-            data_copy["md"] = {}
-            data_copy = populate_data_md_globs(target.target_meta, data_copy)
-            data_copy = populate_data_yaml(target.target_meta, data_copy)
-            data_copy = populate_data_yaml_keyed(target.target_meta, data_copy)
-            data_copy = populate_data_md(target.target_meta, data_copy)
-            if "data_variables" in target.target_meta:
-                data_copy.update(target.target_meta["data_variables"])
-        elif target.root_cfg["version"] >= 1:
-            _LOGGER.info("MM | Processing config version 1...")
-            if "data" in target.target_meta:
-                processed_data_block = process_data_block(
-                    target.target_meta["data"], target.target_meta["_filepath"]
-                )
-            else:
-                processed_data_block = process_data_block(
-                    {}, target.target_meta["_filepath"]
-                )
-            _LOGGER.debug("processed_data_block:", processed_data_block)
-            data_copy.update(processed_data_block)
+        if "version" in tgt.root_cfg and not tgt.root_cfg["version"] >= 1:
+            _LOGGER.error("Can't process this config version.")
+
+        _LOGGER.info("MM | Processing config version 1...")
+        if "data" in tgt.meta:
+            processed_data_block = process_data(tgt.meta["data"], tgt.meta["_filepath"])
+        else:
+            processed_data_block = process_data({}, tgt.meta["_filepath"])
+        _LOGGER.debug("processed_data_block:", processed_data_block)
+        data_copy.update(processed_data_block)
+
         k = list(data_copy.keys())
         _LOGGER.info(f"MM | Available keys: {k}")
         if MD_FILES_KEY in data_copy:
@@ -583,27 +456,35 @@ class MarkdownMelder(object):
             )
         return data_copy
 
-    def render_template(self, melded_input, target, double=True):
+    def render_template(self, melded_input, target, double=None):
         # print(melded_input)
         if "data" not in melded_input:
             melded_input["data"] = {}
-        if "md_template" in target.target_meta:
+        if "md_template" in target.meta:
             _LOGGER.error(
                 "Please update your config! 'md_template' was renamed to 'jinja_template'."
             )
-            target.target_meta["jinja_template"] = target.target_meta["md_template"]
+            target.meta["jinja_template"] = target.meta["md_template"]
 
-        if (
-            "jinja_template" in target.target_meta
-            and target.target_meta["jinja_template"]
-        ):
-            tpl = load_template(target.target_meta)
+        if "jinja_template" in target.meta and target.meta["jinja_template"]:
+            tpl = load_template(target.meta)
         else:
             # cmd_data["jinja_template"] = None
             tpl = Template(tpl_generic)
             _LOGGER.error(
                 "No jinja_template provided. Using generic markmeld jinja_template."
             )
+
+        if double is None:
+            if (
+                "recursive_render" in target.meta
+                and not target.meta["recursive_render"]
+            ):
+                double = False
+            else:
+                # Recursive rendering allows your template to include variables
+                double = True
+
         if double:
             return Template(tpl.render(melded_input)).render(melded_input)  # two times
         else: