+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.0] -- 2022-11-21
+
+### Changed
+
+- Removed support for original beta config version
+- Implemented base targets instead of inheriting from root config
+- Completed refactoring and various bugfixes and tests
+
 ## [0.0.2] -- 2022-11-18
 
 This is a transitional release that can handle both original beta and version 1 config formats.